# api/docs/swagger.yaml
basePath: /
definitions:
  handler.RecognizeResponse:
    properties:
      matched_image:
        type: string
      method:
        description: '"ml" or "hash"'
        type: string
      processing_time_ms:
        type: integer
      result:
        type: string
      similarity:
        type: number
    type: object
info:
  contact: {}
  description: API for image recognition using ML and perceptual hashing
  title: Photo Recognition API
  version: "1.1"
paths:
  /admin/add:
    post:
      consumes:
      - multipart/form-data
      description: Add reference image to database
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      - description: Custom image name
        in: formData
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add new image
      tags:
      - Image Database Management
  /admin/hello:
    get:
      description: Test connection endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hello endpoint
      tags:
      - Image Database Management
  /admin/toggle-ml:
    post:
      consumes:
      - multipart/form-data
      description: Enable/disable ML-based recognition
      parameters:
      - description: Set to 'true' or 'false'
        in: formData
        name: enable
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Toggle ML mode
      tags:
      - Image Database Management
  /recognize:
    post:
      consumes:
      - multipart/form-data
      description: Compare uploaded image against database using ML or hashing
      parameters:
      - description: Image file to check
        in: formData
        name: image
        required: true
        type: file
      - description: Similarity threshold (0-100)
        in: formData
        name: threshold
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RecognizeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recognize image
      tags:
      - Image Recognition
swagger: "2.0"